modularized watchlist details page
